Split some stuff to new module - `signal_routing`
